Change BigDecimal value instead Double value

+++ projectmanagement-api/src/main/java/com/cyacompany/projectmanagement_api/model/Position.java
@@ -3,6 +3,8 @@ package com.cyacompany.projectmanagement_api.model;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "G2M_CARGO")
 @Getter
@@ -19,7 +21,7 @@ public class Position {
   private String description;
 
   @Column(name = "CarCosHor", precision = 6, scale = 2)
-  private Double costPerHour;
+  private BigDecimal costPerHour;
 
   @Column(name = "CarEstReg", length = 1, nullable = false)
   private String status;
